requirements.txt
asgiref==3.5.2
bleach==5.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
django-ckeditor==6.5.1
django-countries==7.3.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-richtextfield==1.6.1
django-summernote==0.8.20.0
idna==3.4
PyMySQL==1.0.2
python-slugify==6.1.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
webencodings==0.5.1
